### 2024-03-14 #6488 @editorialbot
[PRE REVIEW]: terratools: A Python package to analyse TERRA mantle convection simulations
**Submitting author:** @jamespanton93 (<a href="http://orcid.org/0000-0002-1345-9406">James Panton</a>)
**Repository:** https://github.com/mantle-convection-constrained/terratools
**Branch with paper.md** (empty if default branch): Update_JOSS
**Version:** v1.0.2
