### 2024-03-15 439ca06
Add missing DOIs in paper.bib
Added a couple of missing DOIs

Signed-off-by: jamespanton93 <pantonjc@cardiff.ac.uk>
