+++ paper/paper.bib
@@ -24,7 +24,8 @@
   number={5},
   pages={501--511},
   year={1985},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/BF01008348}
 }
 
 @article{bunge1997,
@@ -35,7 +36,8 @@
   number={B6},
   pages={11991--12007},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/96JB03806}
 }
 @article{Davies2011,
   title={Fluidity: A fully unstructured anisotropic adaptive mesh computational modeling framework for geodynamics},
