### 2024-03-15 #6488 @jamespanton93
Hi Kristen,

I have added in the missing DOIs on branch Update_JOSS
Suggest reviewers are: [leouieda](https://github.com/leouieda), [calum-chamberlain](https://github.com/calum-chamberlain), [s-m-e](https://github.com/s-m-e), [TobbeTripitaka](https://github.com/TobbeTripitaka), [Haipeng-ustc](https://github.com/Haipeng-ustc)
