### 2024-06-03 468cb6f
Allow passage of fig and ax handles to plot_layer and plot_section, also return the cbar object to allow manipulation.
