### 2024-06-03 db54037
pass fig and ax handles to model.plot_spectral_heterogeneity and model.plot_hp_map
