### 2024-06-03 9a7f695
create fig and ax handles if not supplied for x-section plot
