### 2024-06-03 ababdbf
Add functionality for plotting a given component of vector fields for `plot_layer`, `plot_section`, `plot_spectral_heterogeneity`, `plot_hp_map`.
