+++ terratools/terra_model.py
@@ -525,7 +525,7 @@ class TerraModel:
         """
         return self._fields.keys()
 
-    def evaluate(self, lon, lat, r, field, method="triangle", depth=False):
+    def evaluate(self, lon, lat, r, field, method="triangle", depth=False, v_field_ind=None):
         """
         Evaluate the value of field at radius r km, latitude lat degrees
         and longitude lon degrees.
@@ -556,6 +556,10 @@ class TerraModel:
         """
         _check_field_name(field)
         self._check_has_field(field)
+        
+        if _is_vector_field(field) and v_field_ind==None:
+            print('v_field_ind not supplied, defaulting to 0')
+            v_field_ind=0
 
         if method not in ("triangle", "nearest"):
             raise ValueError("method must be one of 'triangle' or 'nearest'")
@@ -580,7 +584,7 @@ class TerraModel:
             r = self.to_radius(r)
 
         lons, lats = self.get_lateral_points()
-        array = self.get_field(field)
+        array = (self.get_field(field) if _is_scalar_field(field) else self.get_field(field)[:,:,v_field_ind])
 
         # Find bounding layers
         ilayer1, ilayer2 = _bounding_indices(r, radii)
@@ -1150,7 +1154,7 @@ class TerraModel:
         return self._surface_radius - depth
 
     def calc_spherical_harmonics(
-        self, field, nside=2**6, lmax=16, savemap=False, use_pixel_weights=False
+        self, field, nside=2**6, lmax=16, savemap=False, use_pixel_weights=False, v_field_ind=None
     ):
         """
         Function to calculate spherical harmonic coefficients for given global field.
@@ -1172,8 +1176,12 @@ class TerraModel:
         :param savemap: Default (``False``) do not save the healpix map
         :type savemap: bool
         """
+        
+        if _is_vector_field(field) and v_field_ind==None:
+            print('v_field_ind not supplied, defaulting to 0')
+            v_field_ind=0
 
-        field_values = self.get_field(field)
+        field_values = (self.get_field(field) if _is_scalar_field(field) else self.get_field(field)[:,:,v_field_ind])
 
         lons, lats = self.get_lateral_points()
 
@@ -1202,10 +1210,16 @@ class TerraModel:
             else:
                 hp_ir[r] = {"power_per_l": power_per_l, "coeffs": hp_coeffs}
         try:
-            self._sph[field] = hp_ir
+            if _is_scalar_field(field):
+                self._sph[field] = hp_ir
+            elif _is_vector_field(field):
+                self._sph[f"{field}{v_field_ind}"] = hp_ir
         except:
             self._sph = {}
-            self._sph[field] = hp_ir
+            if _is_scalar_field(field):
+                self._sph[field] = hp_ir
+            elif _is_vector_field(field):
+                self._sph[f"{field}{v_field_ind}"] = hp_ir
 
     def plot_hp_map(
         self,
@@ -1220,6 +1234,7 @@ class TerraModel:
         delta=None,
         extent=(-180, 180, -90, 90),
         method="nearest",
+        v_field_ind=None,
         show=True,
         **subplots_kwargs,
     ):
@@ -1270,6 +1285,9 @@ class TerraModel:
 
         :returns: figure, axis and colourbar handles
         """
+        if _is_vector_field(field) and v_field_ind==None:
+            print('v_field_ind not supplied, defaulting to 0')
+            v_field_ind=0
 
         if radius is None and index is None:
             raise ValueError("Either radius or index must be given")
@@ -1283,12 +1301,16 @@ class TerraModel:
 
             layer_index = index
             layer_radius = radii[index]
-
-        dat = self.get_spherical_harmonics(field)[layer_index]["coeffs"]
[…]
