### 2024-06-03 2d54c2d
Update calls in tests and examples to plotting functions
