+++ terratools/terra_model.py
@@ -1301,7 +1301,15 @@ class TerraModel:
             label = title
 
         fig, ax, cbar = plot.layer_grid(
-            fig, ax, lon, lat, rad, hp_remake, delta=delta, extent=extent, label=label
+            lon,
+            lat,
+            rad,
+            hp_remake,
+            delta=delta,
+            extent=extent,
+            label=label,
+            fig=fig,
+            ax=ax,
         )
 
         if depth:
@@ -1393,8 +1401,8 @@ class TerraModel:
             savepath,
             lyrmin,
             lyrmax,
-            fig,
-            ax,
+            fig=fig,
+            ax=ax,
             **subplots_kwargs,
         )
 
@@ -1481,8 +1489,6 @@ class TerraModel:
             cmap = _FIELD_COLOUR_SCALE[field]
 
         fig, ax, cbar = plot.layer_grid(
-            fig,
-            ax,
             lon,
             lat,
             layer_radius,
@@ -1495,6 +1501,8 @@ class TerraModel:
             cmap=cmap,
             vmin=vmin,
             vmax=vmax,
+            fig=fig,
+            ax=ax,
         )
 
         if depth:
@@ -1640,8 +1648,6 @@ class TerraModel:
             cmap = _FIELD_COLOUR_SCALE[field]
 
         fig, ax, cbar = plot.plot_section(
-            fig,
-            ax,
             distances,
             radii,
             grid,
@@ -1649,6 +1655,8 @@ class TerraModel:
             levels=levels,
             show=show,
             label=label,
+            fig=fig,
+            ax=ax,
         )
 
         return fig, ax, cbar
@@ -1939,9 +1947,7 @@ class TerraModel:
             label = "n-layers plume detected"
             radius = 0.0
 
-            fig, ax = plot.layer_grid(
-                fig,
-                ax,
+            fig, ax, cbar = plot.layer_grid(
                 lon,
                 lat,
                 radius,
@@ -1951,6 +1957,8 @@ class TerraModel:
                 label=label,
                 method=method,
                 coastlines=coastlines,
+                fig=fig,
+                ax=ax,
             )
 
             mindep = np.min(self.plm_depth_range)
@@ -1966,7 +1974,7 @@ class TerraModel:
             if show:
                 fig.show()
 
-            return fig, ax
+            return fig, ax, cbar
 
         def plot_plumes_3d(
             self,
