+++ examples/example_plot_flowfield.py
@@ -66,7 +66,7 @@ geographic vector is defined as positive [east, north, up].
 """
 
 # %%
-fig, ax = plt.subplots()
+fig, ax, cbar = plt.subplots()
 max_zvel = np.nanmax(flow_velocity[2, :, :])
 psm = ax.pcolor(
     glon, glat, flow_velocity[2, :, :], vmin=-max_zvel, vmax=max_zvel, cmap="seismic"
