+++ terratools/terra_model.py
@@ -525,7 +525,9 @@ class TerraModel:
         """
         return self._fields.keys()
 
-    def evaluate(self, lon, lat, r, field, method="triangle", depth=False, v_field_ind=None):
+    def evaluate(
+        self, lon, lat, r, field, method="triangle", depth=False, v_field_ind=None
+    ):
         """
         Evaluate the value of field at radius r km, latitude lat degrees
         and longitude lon degrees.
@@ -556,10 +558,10 @@ class TerraModel:
         """
         _check_field_name(field)
         self._check_has_field(field)
-        
-        if _is_vector_field(field) and v_field_ind==None:
-            print('v_field_ind not supplied, defaulting to 0')
-            v_field_ind=0
+
+        if _is_vector_field(field) and v_field_ind == None:
+            print("v_field_ind not supplied, defaulting to 0")
+            v_field_ind = 0
 
         if method not in ("triangle", "nearest"):
             raise ValueError("method must be one of 'triangle' or 'nearest'")
@@ -584,7 +586,11 @@ class TerraModel:
             r = self.to_radius(r)
 
         lons, lats = self.get_lateral_points()
-        array = (self.get_field(field) if _is_scalar_field(field) else self.get_field(field)[:,:,v_field_ind])
+        array = (
+            self.get_field(field)
+            if _is_scalar_field(field)
+            else self.get_field(field)[:, :, v_field_ind]
+        )
 
         # Find bounding layers
         ilayer1, ilayer2 = _bounding_indices(r, radii)
@@ -1154,7 +1160,13 @@ class TerraModel:
         return self._surface_radius - depth
 
     def calc_spherical_harmonics(
-        self, field, nside=2**6, lmax=16, savemap=False, use_pixel_weights=False, v_field_ind=None
+        self,
+        field,
+        nside=2**6,
+        lmax=16,
+        savemap=False,
+        use_pixel_weights=False,
+        v_field_ind=None,
     ):
         """
         Function to calculate spherical harmonic coefficients for given global field.
@@ -1176,12 +1188,16 @@ class TerraModel:
         :param savemap: Default (``False``) do not save the healpix map
         :type savemap: bool
         """
-        
-        if _is_vector_field(field) and v_field_ind==None:
-            print('v_field_ind not supplied, defaulting to 0')
-            v_field_ind=0
 
-        field_values = (self.get_field(field) if _is_scalar_field(field) else self.get_field(field)[:,:,v_field_ind])
+        if _is_vector_field(field) and v_field_ind == None:
+            print("v_field_ind not supplied, defaulting to 0")
+            v_field_ind = 0
+
+        field_values = (
+            self.get_field(field)
+            if _is_scalar_field(field)
+            else self.get_field(field)[:, :, v_field_ind]
+        )
 
         lons, lats = self.get_lateral_points()
 
@@ -1248,8 +1264,8 @@ class TerraModel:
 
         :param radius: radius to plot (nearest model radius is shown)
         :type radius: float
-        
-        :param depth: interpret radius as depth 
+
+        :param depth: interpret radius as depth
         :type depth: bool
 
         :param fig: figure handle
@@ -1285,9 +1301,9 @@ class TerraModel:
 
         :returns: figure, axis and colourbar handles
         """
-        if _is_vector_field(field) and v_field_ind==None:
-            print('v_field_ind not supplied, defaulting to 0')
-            v_field_ind=0
+        if _is_vector_field(field) and v_field_ind == None:
+            print("v_field_ind not supplied, defaulting to 0")
+            v_field_ind = 0
 
         if radius is None and index is None:
             raise ValueError("Either radius or index must be given")
[…]
