### 2024-06-03 12ac660
correction to allow plotting of vector fields in cross section
