### 2024-06-03 088d8e4
fix to allow calculating spherical harmonics of scalar and vector field components
