### 2024-06-05 36cefcc
passing of kwargs into spherical harmonic spectrum plot
