### 2024-06-06 c3d1181
return cbar from plotting functions only if passed `return_cbar=True`
