+++ terratools/terra_model.py
@@ -1215,6 +1215,7 @@ class TerraModel:
         depth=False,
         fig=None,
         ax=None,
+        return_cbar=False,
         nside=2**6,
         title=None,
         delta=None,
@@ -1300,7 +1301,7 @@ class TerraModel:
         else:
             label = title
 
-        fig, ax, cbar = plot.layer_grid(
+        returned = plot.layer_grid(
             lon,
             lat,
             rad,
@@ -1310,7 +1311,12 @@ class TerraModel:
             label=label,
             fig=fig,
             ax=ax,
+            return_cbar=return_cbar,
         )
+
+        fig = returned[0]
+        ax = returned[1]
+        cbar = returned[2] if return_cbar else None
 
         if depth:
             ax.set_title(f"Depth = {int(layer_radius)} km")
@@ -1320,7 +1326,10 @@ class TerraModel:
         if show:
             fig.show()
 
-        return fig, ax, cbar
+        if return_cbar:
+            return fig, ax, cbar
+        else:
+            return fig, ax
 
     def plot_spectral_heterogeneity(
         self,
@@ -1330,6 +1339,7 @@ class TerraModel:
         savepath=None,
         fig=None,
         ax=None,
+        return_cbar=False,
         lmin=1,
         lmax=None,
         lyrmin=1,
@@ -1391,7 +1401,7 @@ class TerraModel:
         radii = self.get_radii()
         depths = self.get_radii()[-1] - radii
 
-        fig, ax, cbar = plot.spectral_heterogeneity(
+        returned = plot.spectral_heterogeneity(
             powers,
             title,
             depths,
@@ -1403,13 +1413,21 @@ class TerraModel:
             lyrmax,
             fig=fig,
             ax=ax,
+            return_cbar=return_cbar,
             **subplots_kwargs,
         )
+
+        fig = returned[0]
+        ax = returned[1]
+        cbar = returned[2] if return_cbar else None
 
         if show:
             fig.show()
 
-        return fig, ax, cbar
+        if return_cbar:
+            return fig, ax, cbar
+        else:
+            return fig, ax
 
     def calc_bulk_composition(self):
         """
@@ -1435,6 +1453,7 @@ class TerraModel:
         delta=None,
         fig=None,
         ax=None,
+        return_cbar=False,
         extent=(-180, 180, -90, 90),
         method="nearest",
         coastlines=True,
@@ -1488,7 +1507,7 @@ class TerraModel:
         if cmap is None:
             cmap = _FIELD_COLOUR_SCALE[field]
 
-        fig, ax, cbar = plot.layer_grid(
+        returned = plot.layer_grid(
             lon,
             lat,
             layer_radius,
@@ -1503,7 +1522,12 @@ class TerraModel:
             vmax=vmax,
             fig=fig,
             ax=ax,
+            return_cbar=return_cbar,
         )
+
+        fig = returned[0]
+        ax = returned[1]
+        cbar = returned[2] if return_cbar else None
 
         if depth:
             ax.set_title(f"Depth {int(layer_radius)} km")
@@ -1511,7 +1535,10 @@ class TerraModel:
         if show:
             fig.show()
 
-        return fig, ax, cbar
+        if return_cbar:
+            return fig, ax, cbar
+        else:
+            return fig, ax
 
     def plot_section(
         self,
@@ -1522,6 +1549,7 @@ class TerraModel:
         distance,
         fig=None,
         ax=None,
+        return_cbar=False,
         minradius=None,
         maxradius=None,
         delta_distance=1,
@@ -1647,7 +1675,7 @@ class TerraModel:
         if cmap is None:
             cmap = _FIELD_COLOUR_SCALE[field]
 
-        fig, ax, cbar = plot.plot_section(
+        returned = plot.plot_section(
             distances,
[…]
