### 2024-06-06 3c9b0cf
update DOIs and version number
