### 2024-06-10 8fb6826
Changed lambda function for PREM pressure to a named function so model pickling works again.
