### 2024-07-24 b101164
Fix reading of NetCDF files without composition
Previously, NetCDF files read with `terra_model.read_netcdf`
were not assumed to contain composition, but nonetheless trying
to read files without it would cause an error because of an
erroneous check on the number of composition components.  Fix
this so that model files without composition can be read.
