+++ pyproject.toml
@@ -20,14 +20,14 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = "^3.8"
-numpy = "^1.23"
+numpy = ">1.23"
 scipy = "^1.6"
 netcdf4 = "^1"
 matplotlib = "^3.5"
 scikit-learn = "^1.3.0"
 healpy = "^1.16.2"
 pooch = "^1.7"
-
+shapely = "^2"
 
 [tool.poetry.dev-dependencies]
 ipython = "^8.5.0"
