### 2024-10-15 362b2af
first implementation of plume buoyancy flux calculation
