### 2024-10-23 572bf07
add test for buoyancy flux
