### 2024-11-29 #7539 @editorialbot
[REVIEW]: terratools: A Python package to analyse TERRA mantle convection simulations
**Submitting author:** @jamespanton93 (<a href="http://orcid.org/0000-0002-1345-9406">James Panton</a>)
**Repository:** https://github.com/mantle-convection-constrained/terratools
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2.0

@bch0w, @jessepisel, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
