### 2024-12-19 #7539 @bch0w
Hi @jamespanton93 , finally getting around to this review. I am working through install procedures for Terratools and have hit the following error while following both the PyPi version and the development version of the code. Error message is below.

I am working on a Mac M1 Running OSX13.1 and followed the pre-install directions. Some slight deviations from your install instructions are that I am attempting to install into an empty Conda environment to keep things cleaner, and I used `--no-cache-dir` just to check if my cached packages were causing an issue, but the same error arose without that flag option.

Let me know what to try to resolve this, thanks!

```bash
(terratools) ➜  terratools git:(main) ✗ python -m pip install terratools --no-cache-dir
Collecting terratools
  Downloading terratools-1.1.0-py3-none-any.whl.metadata (4.0 kB)
Collecting healpy<2.0.0,>=1.16.2 (from terratools)
  Downloading healpy-1.18.0.tar.gz (4.8 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting matplotlib<4.0,>=3.5 (from terratools)
  Downloading matplotlib-3.10.0-cp313-cp313-macosx_11_0_arm64.whl.metadata (11 kB)
Collecting netcdf4<2,>=1 (from terratools)
  Downloading netcdf4-1.7.2.tar.gz (835 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 835.1/835.1 kB 24.6 MB/s eta 0:00:00
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [32 lines of output]
      -L/opt/homebrew/Cellar/hdf5/1.14.5/lib -lhdf5
      reading from setup.cfg...
      NETCDF4_DIR environment variable not set, checking standard locations..
      checking /opt/miniconda3/envs/terratools...
      checking /opt/miniconda3/envs/terratools/Library...
      checking /Users/chow...
      checking /usr/local...
      checking /sw...
      checking /opt...
      checking /opt/local...
      checking /opt/homebrew...
      checking /usr...
      Traceback (most recent call last):
        File "/opt/miniconda3/envs/terratools/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
          ~~~~^^
        File "/opt/miniconda3/envs/terratools/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
                                   ~~~~^^^^^^^^^^^^^^^^^^^^^^^^
        File "/opt/miniconda3/envs/terratools/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 118, in get_requires_for_build_wheel
          return hook(config_settings)
        File "/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-build-env-k5zdmc0n/overlay/lib/python3.13/site-packages/setuptools/build_meta.py", line 334, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=[])
                 ~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-build-env-k5zdmc0n/overlay/lib/python3.13/site-packages/setuptools/build_meta.py", line 304, in _get_build_requires
          self.run_setup()
          ~~~~~~~~~~~~~~^^
        File "/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-build-env-k5zdmc0n/overlay/lib/python3.13/site-packages/setuptools/build_meta.py", line 320, in run_setup
          exec(code, locals())
          ~~~~^^^^^^^^^^^^^^^^
        File "<string>", line 292, in <module>
      ValueError: did not find netCDF version 4 headers
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

```
