### 2024-12-19 mantle-convection-constrained/terratools#171 @jamespanton93
Hi @bch0w, thanks for getting to this! I have just done an install in a fresh Conda env. I ran `pip install cartopy` followed by `pip install terratools`, but could not recreate your error (Also running on an M1 but with OSX 15.1.1). From the error messages it looks like it could be something to do with your hdf5 homebrew installation interfering with the netcdf4 package? Perhaps you could try uninstalling that and trying from fresh again?

Hopefully I can have a deeper look soon and I'll see whether I can recreate the error.

Cheers,
James
