### 2025-01-15 mantle-convection-constrained/terratools#171 @bch0w
Thanks for the advice, @jamespanton93, turns out this was a me problem. For others who might face the same issue, the solution was to install netCDF and hdf5 via homebrew on my Mac

```bash
brew install hdf5 netcdf
python -m pip install terratools
```

I'll be continuing with my review in: https://github.com/openjournals/joss-reviews/issues/7539
