### 2025-01-15 #7539 @bch0w
Happy new year @observingClouds! Thanks for the check, things had stalled out at installation (https://github.com/mantle-convection-constrained/terratools/issues/171) but I realized this was a me problem and that issue is now closed. I will continue with the review now that I have the package installed!
