### 2025-01-16 #7539 @bch0w
Hi @observingClouds, I've completed my review checklist (https://github.com/openjournals/joss-reviews/issues/7539#issuecomment-2512058356). Kudos to the authors for a nicely presented package!

I've left two items unchecked (installation and tests) on the checklist. There were some manual installation steps I had to take where additional documentation would have been helpful, and when I ran the automated tests I received a few errors.

I left comments in blockquote format where I thought they would help. Happy to provide additional information where necessary.
