### 2025-01-16 mantle-convection-constrained/terratools#171 @anowacki
Hi @bch0w.  Thank you very much for your review and suggestions.

For this issue, just to add for the record in future, you shouldn't need to install anything separately for terratools to work.  I also cannot reproduce this in a fresh conda environment (on Linux or Mac).  I agree with James that it seems possible this is some sort of Homebrew interference, since HDF5 is getting installed, but the installer is picking up your Homebrew HDF5 install when building the package.  If so, it's likely that you have incompatible

Did you install a specific version of Python when you create your environment like below?
```sh
conda create -n terratools python=3.10
```
If not, can you try that?  If so, can you try using an earlier version of `python`?  We have not tested v3.13 much, and I see you are using that.
