### 2025-01-16 65a74e7
Allow use of NumPy v2 and newer versions of Cartopy
Support NumPy v2 but not future major version changes.

Versions of Cartopy from v0.21 to v0.24 do not introduce any
breaking changes affecting us.
