+++ pyproject.toml
@@ -19,8 +19,8 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
-numpy = ">1.23"
+python = "^3.10"
+numpy = "^1.23, ^2"
 scipy = "^1.6"
 netcdf4 = "^1"
 matplotlib = "^3.5"
@@ -35,7 +35,7 @@ black = "^23.9.1"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.9.1"
-cartopy = "^0.21.1"
+cartopy = "^0.21.1, ^0.22, ^0.23, ^0.24"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
