### 2025-01-16 163c541
CI: Update the upload-artifact GitHub action version
Some CI runs are currently failing because we use a version of
`actions/upload-artifact` which is too old
(https://github.blog/changelog/2024-02-13-deprecation-notice-v1-and-v2-of-the-artifact-actions/).
Replace these uses with the latest version to fix the CI runs.

Longer term, updating them automatically may be prudent
(https://github.com/mantle-convection-constrained/terratools/issues/172).
