### 2025-01-16 #7539 @anowacki
Thank you for your very helpful review, @bch0w!

Just to your unticked things:

> - [ ] **Installation:** Does installation proceed as outlined in the documentation?
>> 1. Installation worked but may need some adjustments or additional pre-installation notes. I add the caveat that I was installing this package into a Conda environment to keep my base environment clean, which deviates from the instructions given by the authors. I think the option for virtual environments might be useful for future users, too, but not a requirement.

Installation into conda environments is supported and should work fine (indeed, I would encourage it like you!), so I have added that to `README.md` now (https://github.com/mantle-convection-constrained/terratools/pull/176).

>> 2. NetCDF had to be installed separately through the Mac package manager, Homebrew. (https://github.com/mantle-convection-constrained/terratools/issues/171). It would be great to have that as another pre-installation note in the install instructions in the case that other users face a similar issue.

>> 3. I had to separately install healpy as I was getting an ImportError (similar to
https://github.com/healpy/healpy/issues/846). This was strange as I saw the pip installation install healpy. After the manual install conda install healpy, things worked.

Neither of these things should be required, and as commented in https://github.com/mantle-convection-constrained/terratools/issues/171#issuecomment-2595667824, I think this might be an issue with the Python environment not this package in particular since we haven't been able to reproduce it locally or on CI.  Therefore I don't think adding them to the `README.md` file would help in this particular case.

> distutils has been rolled into setuptools in Python>3.12, so Users using these versions of Python will have to install setuptools to access distutils (this was required by `example_convert.py`).

Thank you very much for alerting us to this!  We have [replaced the use of `distutils` with `shutil`](https://github.com/mantle-convection-constrained/terratools/pull/177).

> - [ ] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>> Dependencies are listed in `terratools/pyproject.toml`, with additional installation requirements not controlled by the package manager (Pip) listed in `terratools/README.md`. I would suggest some additions to these instructions, following what is mentioned in Functionality.

As above, we think the installation issue is a general one, not related to terratools specifically.  More detail on installing Cartopy is now in the `README.md` as noted above.

> - [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
>> Yes there are automated tests, when I ran all tests, two tests failed and three error'd (brief output below, I'm happy to provide the full error output).

Yes, please do open an issue for any failing tests with the full output and details of your system and environment!  That would be very helpful.  These aren't being reproduced by CI if they are real.

>> It might be helpful to include a readme for how to run tests, I did not see anything mentioned in the documentation, but I was able to determine how to run all tests.

This [has been added to the `CONTRIBUTING.md` file](https://github.com/mantle-convection-constrained/terratools/pull/175).

Hopefully these satisfy those requirements and you are able to tick them off after this, but please do continue to open issues about them otherwise.
