### 2025-01-17 mantle-convection-constrained/terratools#179 @bch0w
Failing test_repr (test_terra_model.TestTerraModelRepr)
Related to: https://github.com/openjournals/joss-reviews/issues/7539

Hi Terratools, here is one test that is failing on my local build related to the JOSS review linked above. Relevant build information:
- Platform: Macbook Pro M1 running MacOS 13.1
- Python 3.10.16
- Terratools Commit Hash: f118312

Happy to provide more information as needed!

```sh
(terratools) ➜  terratools git:(main) ✗ python -m unittest discover tests
..................................../Users/chow/Work/software/terratools/terratools/lookup_tables.py:473: UserWarning: One or more of your inputs exceeds the table range, reverting to maximum table range
  warn(
./Users/chow/Work/software/terratools/terratools/lookup_tables.py:478: UserWarning: One or more of your inputs is below the table range, reverting to minimum table range
  warn(
...............adding geographic flow velocities
k-means analysis
density based clustering
Detected 2 plumes with 2308 noise points
...................................................................F..........
FAIL: test_repr (test_terra_model.TestTerraModelRepr)
Traceback (most recent call last):
  File "/Users/chow/Work/software/terratools/tests/test_terra_model.py", line 396, in test_repr
    self.assertEqual(
AssertionError: "Terr[64 chars]ts: (np.float32(1000.0), np.float32(2000.0))\n[179 chars]alse" != "Terr[64 chars]ts: (1000.0, 2000.0)\n  number of lateral poin[155 chars]alse"
  TerraModel:
             number of radii: 3
-              radius limits: (np.float32(1000.0), np.float32(2000.0))
?                             -----------       -  -----------       -
+              radius limits: (1000.0, 2000.0)
    number of lateral points: 3
                      fields: ['t', 'c_hist']
           composition names: ['a', 'b']
          composition values: None
           has lookup tables: False

Ran 130 tests in 6.227s

FAILED (failures=1)
```
