### 2025-01-17 mantle-convection-constrained/terratools#171 @bch0w
Hi @anowacki, thanks for your response! I agree that the HDF5 issue was a local issue as it was affecting some of my other Conda installations. I'll make note of that in the review.

I tried installing with Python 3.10 but I still get the Healpy import issue. This is strange because I see Pip installing and building Healpy, but when I go to import Terratools I get the following import issue copied at the bottom of this comment.

I am able to circumvent this by installing Healpy through Conda `conda install healpy` but understandably that shouldn't be needed. Not sure if this is another local issue if you are unable to recreate but thought I'd point it out to see if we can resolve.

Also happy to open this in a separate issue if it's warranted!

### Installation
```bash
(terratools) ➜  terratools git:(main) ✗ python -m pip install terratools --no-cache-dir
Collecting terratools
  Downloading terratools-1.1.0-py3-none-any.whl.metadata (4.0 kB)
Collecting healpy<2.0.0,>=1.16.2 (from terratools)
  Downloading healpy-1.18.0.tar.gz (4.8 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting matplotlib<4.0,>=3.5 (from terratools)
  Downloading matplotlib-3.10.0-cp310-cp310-macosx_11_0_arm64.whl.metadata (11 kB)
Collecting netcdf4<2,>=1 (from terratools)
  Downloading netcdf4-1.7.2.tar.gz (835 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 835.1/835.1 kB 37.4 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting numpy<2.0,>=1.23 (from terratools)
  Downloading numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl.metadata (61 kB)
Collecting pooch<2.0,>=1.7 (from terratools)
  Downloading pooch-1.8.2-py3-none-any.whl.metadata (10 kB)
Collecting scikit-learn<2.0.0,>=1.3.0 (from terratools)
  Downloading scikit_learn-1.6.1-cp310-cp310-macosx_12_0_arm64.whl.metadata (31 kB)
Collecting scipy<2.0,>=1.6 (from terratools)
  Downloading scipy-1.15.1-cp310-cp310-macosx_12_0_arm64.whl.metadata (61 kB)
Collecting astropy (from healpy<2.0.0,>=1.16.2->terratools)
  Downloading astropy-6.1.7-cp310-cp310-macosx_11_0_arm64.whl.metadata (10 kB)
Collecting contourpy>=1.0.1 (from matplotlib<4.0,>=3.5->terratools)
  Downloading contourpy-1.3.1-cp310-cp310-macosx_11_0_arm64.whl.metadata (5.4 kB)
Collecting cycler>=0.10 (from matplotlib<4.0,>=3.5->terratools)
  Downloading cycler-0.12.1-py3-none-any.whl.metadata (3.8 kB)
Collecting fonttools>=4.22.0 (from matplotlib<4.0,>=3.5->terratools)
  Downloading fonttools-4.55.3-cp310-cp310-macosx_10_9_universal2.whl.metadata (165 kB)
Collecting kiwisolver>=1.3.1 (from matplotlib<4.0,>=3.5->terratools)
  Downloading kiwisolver-1.4.8-cp310-cp310-macosx_11_0_arm64.whl.metadata (6.2 kB)
Collecting packaging>=20.0 (from matplotlib<4.0,>=3.5->terratools)
  Downloading packaging-24.2-py3-none-any.whl.metadata (3.2 kB)
Collecting pillow>=8 (from matplotlib<4.0,>=3.5->terratools)
  Downloading pillow-11.1.0-cp310-cp310-macosx_11_0_arm64.whl.metadata (9.1 kB)
Collecting pyparsing>=2.3.1 (from matplotlib<4.0,>=3.5->terratools)
  Downloading pyparsing-3.2.1-py3-none-any.whl.metadata (5.0 kB)
Collecting python-dateutil>=2.7 (from matplotlib<4.0,>=3.5->terratools)
  Downloading python_dateutil-2.9.0.post0-py2.py3-none-any.whl.metadata (8.4 kB)
Collecting cftime (from netcdf4<2,>=1->terratools)
  Downloading cftime-1.6.4.post1-cp310-cp310-macosx_11_0_arm64.whl.metadata (8.7 kB)
Collecting certifi (from netcdf4<2,>=1->terratools)
  Downloading certifi-2024.12.14-py3-none-any.whl.metadata (2.3 kB)
Collecting platformdirs>=2.5.0 (from pooch<2.0,>=1.7->terratools)
  Downloading platformdirs-4.3.6-py3-none-any.whl.metadata (11 kB)
Collecting requests>=2.19.0 (from pooch<2.0,>=1.7->terratools)
  Downloading requests-2.32.3-py3-none-any.whl.metadata (4.6 kB)
Collecting joblib>=1.2.0 (from scikit-learn<2.0.0,>=1.3.0->terratools)
  Downloading joblib-1.4.2-py3-none-any.whl.metadata (5.4 kB)
Collecting threadpoolctl>=3.1.0 (from scikit-learn<2.0.0,>=1.3.0->terratools)
  Downloading threadpoolctl-3.5.0-py3-none-any.whl.metadata (13 kB)
Collecting six>=1.5 (from python-dateutil>=2.7->matplotlib<4.0,>=3.5->terratools)
  Downloading six-1.17.0-py2.py3-none-any.whl.metadata (1.7 kB)
Collecting charset-normalizer<4,>=2 (from requests>=2.19.0->pooch<2.0,>=1.7->terratools)
  Downloading charset_normalizer-3.4.1-cp310-cp310-macosx_10_9_universal2.whl.metadata (35 kB)
Collecting idna<4,>=2.5 (from requests>=2.19.0->pooch<2.0,>=1.7->terratools)
  Downloading idna-3.10-py3-none-any.whl.metadata (10 kB)
Collecting urllib3<3,>=1.21.1 (from requests>=2.19.0->pooch<2.0,>=1.7->terratools)
  Downloading urllib3-2.3.0-py3-none-any.whl.metadata (6.5 kB)
Collecting pyerfa>=2.0.1.1 (from astropy->healpy<2.0.0,>=1.16.2->terratools)
  Downloading pyerfa-2.0.1.5-cp39-abi3-macosx_11_0_arm64.whl.metadata (5.7 kB)
Collecting astropy-iers-data>=0.2024.10.28.0.34.7 (from astropy->healpy<2.0.0,>=1.16.2->terratools)
  Downloading astropy_iers_data-0.2025.1.13.0.34.51-py3-none-any.whl.metadata (5.1 kB)
Collecting PyYAML>=3.13 (from astropy->healpy<2.0.0,>=1.16.2->terratools)
  Downloading PyYAML-6.0.2-cp310-cp310-macosx_11_0_arm64.whl.metadata (2.1 kB)
Downloading terratools-1.1.0-py3-none-any.whl (114 kB)
Downloading matplotlib-3.10.0-cp310-cp310-macosx_11_0_arm64.whl (8.0 MB)
Downloading numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl (14.0 MB)
Downloading pooch-1.8.2-py3-none-any.whl (64 kB)
Downloading scikit_learn-1.6.1-cp310-cp310-macosx_12_0_arm64.whl (11.1 MB)
Downloading scipy-1.15.1-cp310-cp310-macosx_12_0_arm64.whl (32.5 MB)
Downloading contourpy-1.3.1-cp310-cp310-macosx_11_0_arm64.whl (253 kB)
Downloading cycler-0.12.1-py3-none-any.whl (8.3 kB)
Downloading fonttools-4.55.3-cp310-cp310-macosx_10_9_universal2.whl (2.8 MB)
Downloading joblib-1.4.2-py3-none-any.whl (301 kB)
Downloading kiwisolver-1.4.8-cp310-cp310-macosx_11_0_arm64.whl (65 kB)
Downloading packaging-24.2-py3-none-any.whl (65 kB)
Downloading pillow-11.1.0-cp310-cp310-macosx_11_0_arm64.whl (3.1 MB)
Downloading platformdirs-4.3.6-py3-none-any.whl (18 kB)
Downloading pyparsing-3.2.1-py3-none-any.whl (107 kB)
Downloading python_dateutil-2.9.0.post0-py2.py3-none-any.whl (229 kB)
Downloading requests-2.32.3-py3-none-any.whl (64 kB)
Downloading certifi-2024.12.14-py3-none-any.whl (164 kB)
Downloading threadpoolctl-3.5.0-py3-none-any.whl (18 kB)
Downloading astropy-6.1.7-cp310-cp310-macosx_11_0_arm64.whl (6.4 MB)
Downloading cftime-1.6.4.post1-cp310-cp310-macosx_11_0_arm64.whl (213 kB)
Downloading astropy_iers_data-0.2025.1.13.0.34.51-py3-none-any.whl (1.9 MB)
Downloading charset_normalizer-3.4.1-cp310-cp310-macosx_10_9_universal2.whl (198 kB)
Downloading idna-3.10-py3-none-any.whl (70 kB)
Downloading pyerfa-2.0.1.5-cp39-abi3-macosx_11_0_arm64.whl (329 kB)
Downloading PyYAML-6.0.2-cp310-cp310-macosx_11_0_arm64.whl (171 kB)
Downloading six-1.17.0-py2.py3-none-any.whl (11 kB)
Downloading urllib3-2.3.0-py3-none-any.whl (128 kB)
Building wheels for collected packages: healpy, netcdf4
  Building wheel for healpy (pyproject.toml) ... done
  Created wheel for healpy: filename=healpy-1.18.0-cp310-cp310-macosx_11_0_arm64.whl size=1505756 sha256=43792a994e1bd4fdd9d6844be15f19f48ae1b72e2fb5cd3bfcbb7c03be103e91
  Stored in directory: /private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-ephem-wheel-cache-kvf3o80r/wheels/3b/80/c6/41a9f73c58aac67d4b3e5b0f9015ee5655c3386c60d1379d9d
  Building wheel for netcdf4 (pyproject.toml) ... done
  Created wheel for netcdf4: filename=netCDF4-1.7.2-cp310-cp310-macosx_11_0_arm64.whl size=607745 sha256=7ce8e7d6082f5e589795618512c1be45cb5fb0a44c66f22a35e7be36a4f573d7
  Stored in directory: /private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-ephem-wheel-cache-kvf3o80r/wheels/9b/0c/58/c9f5a1e9735bbef914dde2b9883f794ef51aee7c8ddec4eac7
Successfully built healpy netcdf4
Installing collected packages: urllib3, threadpoolctl, six, PyYAML, pyparsing, platformdirs, pillow, packaging, numpy, kiwisolver, joblib, idna, fonttools, cycler, charset-normalizer, certifi, astropy-iers-data, scipy, requests, python-dateutil, pyerfa, contourpy, cftime, scikit-learn, pooch, netcdf4, matplotlib, astropy, healpy, terratools
Successfully installed PyYAML-6.0.2 astropy-6.1.7 astropy-iers-data-0.2025.1.13.0.34.51 certifi-2024.12.14 cftime-1.6.4.post1 charset-normalizer-3.4.1 contourpy-1.3.1 cycler-0.12.1 fonttools-4.55.3 healpy-1.18.0 idna-3.10 joblib-1.4.2 kiwisolver-1.4.8 matplotlib-3.10.0 netcdf4-1.7.2 numpy-1.26.4 packaging-24.2 pillow-11.1.0 platformdirs-4.3.6 pooch-1.8.2 pyerfa-2.0.1.5 pyparsing-3.2.1 python-dateutil-2.9.0.post0 requests-2.32.3 scikit-learn-1.6.1 scipy-1.15.1 six-1.17.0 terratools-1.1.0 threadpoolctl-3.5.0 urllib3-2.3.0
```

### Healpy ImportError
```bash
(terratools) ➜  terratools git:(main) ✗ python -c "import terratools"
Warning: cannot import _healpy_pixel_lib module
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/Users/chow/Work/software/terratools/terratools/__init__.py", line 9, in <module>
    from . import terra_model
  File "/Users/chow/Work/software/terratools/terratools/terra_model.py", line 7, in <module>
    import healpy as hp
  File "/opt/miniconda3/envs/terratools/lib/python3.10/site-packages/healpy/__init__.py", line 65, in <module>
    from .sphtfunc import (
  File "/opt/miniconda3/envs/terratools/lib/python3.10/site-packages/healpy/sphtfunc.py", line 32, in <module>
    from . import _healpy_sph_transform_lib as sphtlib
ImportError: dlopen(/opt/miniconda3/envs/terratools/lib/python3.10/site-packages/healpy/_healpy_sph_transform_lib.cpython-310-darwin.so, 0x0002): Library not loaded: /private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-be127ymi/healpy_e686217852a84d2394292646e822e4dc/build/temp.macosx-11.0-arm64-cpython-310/lib/libcfitsio.10.dylib
  Referenced from: <5655E7DC-34A6-35F2-8E08-3C3A36C6991E> /opt/miniconda3/envs/terratools/lib/python3.10/site-packages/healpy/_healpy_sph_transform_lib.cpython-310-darwin.so
  Reason: tried: '/opt/homebrew/opt/geos/lib//libcfitsio.10.dylib' (no such file), '/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-be127ymi/healpy_e686217852a84d2394292646e822e4dc/build/temp.macosx-11.0-arm64-cpython-310/lib/libcfitsio.10.dylib' (no such file), '/System/Volumes/Preboot/Cryptexes/OS/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-be127ymi/healpy_e686217852a84d2394292646e822e4dc/build/temp.macosx-11.0-arm64-cpython-310/lib/libcfitsio.10.dylib' (no such file), '/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-be127ymi/healpy_e686217852a84d2394292646e822e4dc/build/temp.macosx-11.0-arm64-cpython-310/lib/libcfitsio.10.dylib' (no such file), '/usr/local/lib/libcfitsio.10.dylib' (no such file), '/usr/lib/libcfitsio.10.dylib' (no such file, not in dyld cache)
```
