### 2025-01-17 #7539 @bch0w
Hi @anowacki, thanks for replying to my review and for addressing some of the points raised!

> Neither of these things should be required, and as commented in https://github.com/mantle-convection-constrained/terratools/issues/171#issuecomment-2595667824, I think this might be an issue with the Python environment not this package in particular since we haven't been able to reproduce it locally or on CI. Therefore I don't think adding them to the README.md file would help in this particular case.

Regarding HDF5, yep that was a me issue as it happened for another environment creation. Agreed that shouldn't be needed in the README.

Regarding Healpy I still see this issue and would like to get to the bottom of it before checking this one off. I started a new issue here: https://github.com/mantle-convection-constrained/terratools/issues/180

> Yes, please do open an issue for any failing tests with the full output and details of your system and environment! That would be very helpful. These aren't being reproduced by CI if they are real.

Okay! I opened a new issue (https://github.com/mantle-convection-constrained/terratools/issues/179) for the one test failure I am running into. I can no longer reproduce the other failures/errors I saw since I've re-built my Conda environment a few times since then.

Thanks!
