### 2025-01-17 mantle-convection-constrained/terratools#179 @anowacki
Thanks for the report!  Can I ask if you're able to give the versions of your dependencies?  If you're using `conda` I think that's `conda list`.
