### 2025-01-18 mantle-convection-constrained/terratools#179 @anowacki
I have tried to reproduce this on NumPy v1.26 but don't get the same test failure.  However, I think #181 should fix this.  If you are able to try with that branch, that would be great.
