### 2025-01-18 50057ca
CI: Try and test on macOS as well
