+++ .github/workflows/main.yml
@@ -34,11 +34,14 @@ jobs:
 
   test:
     name: test
-    runs-on: [ubuntu-latest]
+    runs-on: ${{ matrix.os }}
 
     strategy:
       matrix:
         python-versions: ['3.9', '3.10', '3.11', '3.12']
+        os:
+          - ubuntu-latest
+          - macos-latest
 
     steps:
     - uses: actions/checkout@v5
@@ -78,11 +81,14 @@ jobs:
 
   test_cartopy:
     name: Test no Cartopy
-    runs-on: [ubuntu-latest]
+    runs-on: ${{ matrix.os }}
 
     strategy:
       matrix:
         python-versions: ['3.9', '3.10', '3.11', '3.12']
+        os:
+          - ubuntu-latest
+          - macos-latest
 
     steps:
     - uses: actions/checkout@v5
