### 2025-01-18 mantle-convection-constrained/terratools#180 @anowacki
> Stored in directory: /private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-ephem-wheel-cache-kvf3o80r/wheels/3b/80/c6/41a9f73c58aac67d4b3e5b0f9015ee5655c3386c60d1379d9d

This seems like `pip` is installing the library into a temporary directory, whereas I think it should be installing it into a permanent place.  I think the difference here is that where I have tested, healpy binaries are downloaded, whereas on your machine the wheels for `netcdf` and `healpy` get built locally (which is unexpected).  I am not a Python expert so at this point I'm just guessing, but if you are able to try from scratch without setting `DYLD_LIBRARY_PATH` or `LD_LIBRARY_PATH` and without any Homebrew binaries in your `PATH`, that might help diagnose.
