### 2025-01-30 mantle-convection-constrained/terratools#179 @bch0w
Okay! I just installed everything on my Linux machine running CentOS 7 and things worked as expected, here is the output from the tests and the conda list just for reference. I have a feeling there are some weird system inconsistencies on my Macbook that were causing some hang ups. Closing as complete as (I think) those were User errors and not package ones!

```sh
(terratools) bhchow@grebe [terratools] $ python -m unittest discover tests
..................................../scratch/bhchow/terratools/terratools/lookup_tables.py:473: UserWarning: One or more of your inputs exceeds the table range, reverting to maximum table range
  warn(
./scratch/bhchow/terratools/terratools/lookup_tables.py:478: UserWarning: One or more of your inputs is below the table range, reverting to minimum table range
  warn(
/scratch/bhchow/terratools/terratools/lookup_tables.py:473: UserWarning: One or more of your inputs exceeds the table range, reverting to maximum table range
  warn(
..............layer_grid require cartopy to be installed.adding geographic flow velocities
k-means analysis
density based clustering
Detected 0 plumes with 5697 noise points
.........................................................................
Ran 125 tests in 4.321s

OK
```

```sh
(terratools) bhchow@grebe [terratools] $ conda list
# packages in environment at /home/bhchow/miniconda3/envs/terratools:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                 conda_forge    conda-forge
_openmp_mutex             4.5                       2_gnu    conda-forge
astropy                   7.0.0                    pypi_0    pypi
astropy-iers-data         0.2025.1.27.0.32.44          pypi_0    pypi
bzip2                     1.0.8                h4bc722e_7    conda-forge
ca-certificates           2024.12.14           hbcca054_0    conda-forge
certifi                   2024.12.14               pypi_0    pypi
cftime                    1.6.4.post1              pypi_0    pypi
charset-normalizer        3.4.1                    pypi_0    pypi
contourpy                 1.3.1                    pypi_0    pypi
cycler                    0.12.1                   pypi_0    pypi
fonttools                 4.55.8                   pypi_0    pypi
healpy                    1.18.0                   pypi_0    pypi
idna                      3.10                     pypi_0    pypi
joblib                    1.4.2                    pypi_0    pypi
kiwisolver                1.4.8                    pypi_0    pypi
ld_impl_linux-64          2.43                 h712a8e2_2    conda-forge
libexpat                  2.6.4                h5888daf_0    conda-forge
libffi                    3.4.2                h7f98852_5    conda-forge
libgcc                    14.2.0               h77fa898_1    conda-forge
libgcc-ng                 14.2.0               h69a702a_1    conda-forge
libgomp                   14.2.0               h77fa898_1    conda-forge
liblzma                   5.6.3                hb9d3cd8_1    conda-forge
libnsl                    2.0.1                hd590300_0    conda-forge
libsqlite                 3.48.0               hee588c1_1    conda-forge
libuuid                   2.38.1               h0b41bf4_0    conda-forge
libxcrypt                 4.4.36               hd590300_1    conda-forge
libzlib                   1.3.1                hb9d3cd8_2    conda-forge
matplotlib                3.10.0                   pypi_0    pypi
ncurses                   6.5                  h2d0b736_3    conda-forge
netcdf4                   1.7.2                    pypi_0    pypi
numpy                     1.26.4                   pypi_0    pypi
openssl                   3.4.0                h7b32b05_1    conda-forge
packaging                 24.2                     pypi_0    pypi
pillow                    11.1.0                   pypi_0    pypi
pip                       25.0               pyh8b19718_0    conda-forge
platformdirs              4.3.6                    pypi_0    pypi
pooch                     1.8.2                    pypi_0    pypi
pyerfa                    2.0.1.5                  pypi_0    pypi
pyparsing                 3.2.1                    pypi_0    pypi
python                    3.11.11         h9e4cc4f_1_cpython    conda-forge
python-dateutil           2.9.0.post0              pypi_0    pypi
pyyaml                    6.0.2                    pypi_0    pypi
readline                  8.2                  h8228510_1    conda-forge
requests                  2.32.3                   pypi_0    pypi
scikit-learn              1.6.1                    pypi_0    pypi
scipy                     1.15.1                   pypi_0    pypi
setuptools                75.8.0             pyhff2d567_0    conda-forge
six                       1.17.0                   pypi_0    pypi
terratools                1.1.0                    pypi_0    pypi
threadpoolctl             3.5.0                    pypi_0    pypi
tk                        8.6.13          noxft_h4845f30_101    conda-forge
tzdata                    2025a                h78e105d_0    conda-forge
urllib3                   2.3.0                    pypi_0    pypi
wheel                     0.45.1             pyhd8ed1ab_1    conda-forge
```
