### 2025-01-30 mantle-convection-constrained/terratools#180 @bch0w
Thanks @anowacki! Sorry for the radio silence here. I re-ran everything on my Linux machine and the environment building and tests (#179) worked as expected, so I am leaning towards a User (me!) issue with some weird discrepancies with either Homebrew, or potentially Mac version incompatibility.

Just incase this is useful for future Users, on the Mac removing anything Homebrew from `PATH` got rid of the private folder building (see first block below), but still ran into the healpy import issue (see second block below, probably same as the original issue post). Conda installing Healpy worked, so I wonder if Conda is pointing towards a different binary than Pip? I still think it may be related to https://github.com/healpy/healpy/issues/846 where the solution was also to install from Conda.

Pip is installing from wheel: `healpy-1.18.0-cp311-cp311-macosx_11_0_arm64.whl`
Conda is installing `conda-forge/osx-arm64::healpy-1.18.0-py311heb8ac0d_2`

Anyways, I don't want to hold up the review any longer as this is getting into the weeds and does not seem to be a Terratools issue since my Linux built just fine. So I'll close this issue and check off the install box for the JOSS review. Thanks for your help! Sorry to drag you into this weird corner of Python! But happy to continue looking into this if you feel it's important.

```sh
Building wheels for collected packages: healpy, netcdf4
  Building wheel for healpy (pyproject.toml) ... done
  Created wheel for healpy: filename=healpy-1.18.0-cp311-cp311-macosx_11_0_arm64.whl size=1505295 sha256=5a38932105f88bf3f3491682cb8cb72607500676365439a4183f94dfca64bab8
  Stored in directory: /Users/chow/Library/Caches/pip/wheels/21/47/b4/abb114679a79610c86ee27430f5d97ac1494039457cd498d91
  Building wheel for netcdf4 (pyproject.toml) ... done
  Created wheel for netcdf4: filename=netCDF4-1.7.2-cp311-cp311-macosx_11_0_arm64.whl size=610314 sha256=232ce8d7c3d6b1e906e5db373bc0e1364ea37e230b35433c2c98585cc2fc5551
  Stored in directory: /Users/chow/Library/Caches/pip/wheels/35/05/c9/671ac7f70347ac9cb96914ee9920b2466c639ef223fbcee4f4
Successfully built healpy netcdf4
Installing collected packages: urllib3, threadpoolctl, six, PyYAML, pyparsing, platformdirs, pillow, packaging, numpy, kiwisolver, joblib, idna, fonttools, cycler, charset-normalizer, certifi, astropy-iers-data, scipy, requests, python-dateutil, pyerfa, contourpy, cftime, scikit-learn, pooch, netcdf4, matplotlib, astropy, healpy, terratools
```

```sh
(terratools) ➜  python
Python 3.11.11 | packaged by conda-forge | (main, Dec  5 2024, 14:21:42) [Clang 18.1.8 ] on darwin
>>> import healpy
Warning: cannot import _healpy_pixel_lib module
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/opt/miniconda3/envs/terratools/lib/python3.11/site-packages/healpy/__init__.py", line 65, in <module>
    from .sphtfunc import (
  File "/opt/miniconda3/envs/terratools/lib/python3.11/site-packages/healpy/sphtfunc.py", line 32, in <module>
    from . import _healpy_sph_transform_lib as sphtlib
ImportError: dlopen(/opt/miniconda3/envs/terratools/lib/python3.11/site-packages/healpy/_healpy_sph_transform_lib.cpython-311-darwin.so, 0x0002): Library not loaded: /private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-99q5mx5o/healpy_e842159e356441f28c4f7335b87868e7/build/temp.macosx-11.0-arm64-cpython-311/lib/libcfitsio.10.dylib
  Referenced from: <7148846C-D77A-3F63-9087-239CDB1B7AC4> /opt/miniconda3/envs/terratools/lib/python3.11/site-packages/healpy/_healpy_sph_transform_lib.cpython-311-darwin.so
  Reason: tried: '/opt/homebrew/opt/geos/lib/libcfitsio.10.dylib' (no such file), '/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-99q5mx5o/healpy_e842159e356441f28c4f7335b87868e7/build/temp.macosx-11.0-arm64-cpython-311/lib/libcfitsio.10.dylib' (no such file), '/System/Volumes/Preboot/Cryptexes/OS/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-99q5mx5o/healpy_e842159e356441f28c4f7335b87868e7/build/temp.macosx-11.0-arm64-cpython-311/lib/libcfitsio.10.dylib' (no such file), '/private/var/folders/2t/_05c_j893kg026n_00m49q4h0000gn/T/pip-install-99q5mx5o/healpy_e842159e356441f28c4f7335b87868e7/build/temp.macosx-11.0-arm64-cpython-311/lib/libcfitsio.10.dylib' (no such file), '/usr/local/lib/libcfitsio.10.dylib' (no such file), '/usr/lib/libcfitsio.10.dylib' (no such file, not in dyld cache)
```
