### 2025-01-31 mantle-convection-constrained/terratools#180 @anowacki
Thanks for checking back in on this, @bch0w.  Closing the issue feels right, but we will need to monitor whether this impacts others in future.
