### 2025-08-22 5f2e0ec
fix: correct parentheses in adiabat calculation. Update test to reflect intended semantics; add regression coverage.
