### 2025-09-05 610286e
CI: Update runner to ubuntu-latest
Previously we specified `ubuntu-20.04` as the runner version, but
as of now (2025-09), this is no longer a supported runner.

Use `ubuntu-latest` instead.
