+++ .github/workflows/main.yml
@@ -13,7 +13,7 @@ jobs:
         python-versions: ['3.10']
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v5
     - name: Set up Python ${{ matrix.python-versions }}
       uses: actions/setup-python@v2
       with:
@@ -41,7 +41,7 @@ jobs:
         python-versions: ['3.8', '3.9', '3.10', '3.11', '3.12']
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v5
     - name: Set up Python ${{ matrix.python-versions }}
       uses: actions/setup-python@v2
       with:
@@ -67,7 +67,7 @@ jobs:
         python-versions: ['3.8', '3.9', '3.10', '3.11', '3.12']
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v5
     - name: Set up Python ${{ matrix.python-versions }}
       uses: actions/setup-python@v2
       with:
