### 2025-09-05 4b1d6aa
Update version to v1.2.0
