### 2025-09-05 6dfae2d
Bump minimum Python version to v3.9
Closes #174.
