### 2025-09-05 6a47bc9
Bump minimum Python version to v3.9
Closes #174.
