+++ .github/workflows/main.yml
@@ -38,7 +38,7 @@ jobs:
 
     strategy:
       matrix:
-        python-versions: ['3.8', '3.9', '3.10', '3.11', '3.12']
+        python-versions: ['3.9', '3.10', '3.11', '3.12']
 
     steps:
     - uses: actions/checkout@v5
@@ -64,7 +64,7 @@ jobs:
 
     strategy:
       matrix:
-        python-versions: ['3.8', '3.9', '3.10', '3.11', '3.12']
+        python-versions: ['3.9', '3.10', '3.11', '3.12']
 
     steps:
     - uses: actions/checkout@v5
