### 2025-09-15 8d84513
Add a hard coded pooch registry
We need to avoid reading this from Figshare, as that seems to fail
in Github actions too often to be useful.
