### 2025-09-22 5776a02
Update changelog, README and DOI for new release
- Mention new minimum Python version in changelog.txt
- Update README.md to reflect the above
- Update DOIs to point to the default DOI for TerraTools as a whole,
  rather than specific releases.
