### 2025-09-30 #7539 @kthyng
@busstoptaktik No pressure if you aren't interested in reviewing, but just in case there is a point of confusion — what to review is stated in a checklist and you don't need to provide any content on your own. You essentially would install the software, make sure the docs make sense, make sure the submission has all necessary basics.
