### 2025-10-27 mantle-convection-constrained/terratools#200 @jessepisel
Python 3.8 requirement
Howdy all :wave:, cool package you have built with TerraTools. I am working through my review checklist https://github.com/openjournals/joss-reviews/issues/7539#issuecomment-3422823578 and wanted to recommend dropping the `Python 3.8` example from the `README` installation instructions. Python 3.8 [is past and at end of life](https://devguide.python.org/versions/) (and 3.9 this month) and removing it from the install instructions would improve install clarity. The `3.11` install instructions should be perfect to get anyone from `3.9` and newer up and running.
