+++ .github/workflows/main.yml
@@ -13,7 +13,7 @@ jobs:
         python-versions: ['3.10']
 
     steps:
-    - uses: actions/checkout@v5
+    - uses: actions/checkout@v6
     - name: Set up Python ${{ matrix.python-versions }}
       uses: actions/setup-python@v6
       with:
@@ -41,7 +41,7 @@ jobs:
         python-versions: ['3.9', '3.10', '3.11', '3.12']
 
     steps:
-    - uses: actions/checkout@v5
+    - uses: actions/checkout@v6
     - name: Set up Python ${{ matrix.python-versions }}
       uses: actions/setup-python@v6
       with:
@@ -85,7 +85,7 @@ jobs:
         python-versions: ['3.9', '3.10', '3.11', '3.12']
 
     steps:
-    - uses: actions/checkout@v5
+    - uses: actions/checkout@v6
     - name: Set up Python ${{ matrix.python-versions }}
       uses: actions/setup-python@v6
       with:
