### 2025-12-04 a21e6d7
JOSS: Add missing NERC grant numbers
The original JOSS paper draft only listed Bristol's grant code
for MC^2.  Add in those for Cardiff and Leeds as well.
