### 2025-12-08 78cc1f6
JOSS: Fix typos in paper
Fix a couple of typographical errors in the JOSS manuscript.
