### 2025-12-12 a8208bc
JOSS: Add missing DOIs
Add missing DOIs to the JOSS paper for Baumgardner et al. (1985)
and Bunge et al. (1997).
