### 2025-12-12 bde58d7
JOSS: Fix Tackley (2008) reference in paper
The cite command for Tackley (2008) was wrong, making the first
citation render incorrectly in the PDF.
