+++ paper/paper.md
@@ -44,7 +44,7 @@ bibliography: paper.bib
 
 # Summary
 
-Fluid-like convection of Earth’s rocky mantle drives processes such as plate tectonics that shape the surface and explains the evolution of our planet on the longest time scales. Because of this, computer simulations of mantle convection have become important for our understanding of the Earth and large scale simulation codes have been created by the community [@Tackley@2008; @Kronbichler2012; @Davies2011; @Moresi2014; @Zhong2000]. One example is TERRA [@baumgardner1985; @bunge1997], a large parallel program using the finite element method to simulate convection.  TERRA is written in Fortran and runs on supercomputers, producing gigabytes of files for each timestep. Handling these outputs is non-trivial because the output reflects the structured grid of finite element mesh and the parallel decomposition used to execute the code. Furthermore, existing closed-source tools in Fortran and other compiled languages present a significant barrier to further development of model analysis tools. Here we describe terratools, a Python package designed to enable reproducible and repeatable post-processing analysis of the outputs of TERRA simulations. Documentation is available via a dedicated website (https://terratools.readthedocs.io/).
+Fluid-like convection of Earth’s rocky mantle drives processes such as plate tectonics that shape the surface and explains the evolution of our planet on the longest time scales. Because of this, computer simulations of mantle convection have become important for our understanding of the Earth and large scale simulation codes have been created by the community [@Tackley2008; @Kronbichler2012; @Davies2011; @Moresi2014; @Zhong2000]. One example is TERRA [@baumgardner1985; @bunge1997], a large parallel program using the finite element method to simulate convection.  TERRA is written in Fortran and runs on supercomputers, producing gigabytes of files for each timestep. Handling these outputs is non-trivial because the output reflects the structured grid of finite element mesh and the parallel decomposition used to execute the code. Furthermore, existing closed-source tools in Fortran and other compiled languages present a significant barrier to further development of model analysis tools. Here we describe terratools, a Python package designed to enable reproducible and repeatable post-processing analysis of the outputs of TERRA simulations. Documentation is available via a dedicated website (https://terratools.readthedocs.io/).
 
 # Statement of need
 
