### 2025-12-12 e1682ee
JOSS: Fix reference formatting
Fix some incorrect capitalisation of references by properly
protecting characters in the .bib file.
