+++ paper/paper.bib
@@ -17,7 +17,7 @@
 }
 
 @article{baumgardner1985,
-  title={Three-dimensional treatment of convective flow in the Earth's mantle},
+  title={Three-dimensional treatment of convective flow in the {E}arth's mantle},
   author={Baumgardner, John R},
   journal={Journal of Statistical Physics},
   volume={39},
@@ -54,7 +54,7 @@
 }
 
 @article{Ghelichkhan2021,
-        title = {Global mantle flow retrodictions for the early {Cenozoic} using an adjoint method: evolving dynamic topographies, deep mantle structures, flow trajectories and sublithospheric stresses},
+        title = {Global mantle flow retrodictions for the early {C}enozoic using an adjoint method: evolving dynamic topographies, deep mantle structures, flow trajectories and sublithospheric stresses},
         volume = {226},
         issn = {0956-540X},
         url = {https://academic.oup.com/gji/article/226/2/1432/6177682},
@@ -95,7 +95,7 @@
   url     = {http://dx.doi.org/10.1111/j.1365-246X.2012.05609.x}
 }
 @software{Moresi2014,
-  title={CitcomS v3.3.1},
+  title={{CitcomS v3.3.1}},
   author={Moresi, Louis and Zhong, Shijie and Han, Lijie and Conrad, Clint and Tan, Eh and Gurnis, Michael and Choi, Eunseo and Thoutireddy, Pururav and Manea, Vlad and McNamara, Allen and Becker, Thorsten and Leng, Wei and Armendariz, Luis},
   year ={2014},
   publisher={Zenodo},
@@ -104,7 +104,7 @@
   url = {https://doi.org/10.5281/zenodo.7271920}
 }
 @article{panton2023,
-        title = {The {Stability} of {Dense} {Oceanic} {Crust} {Near} the {Core}-{Mantle} {Boundary}},
+        title = {The Stability of Dense Oceanic Crust Near the Core-Mantle Boundary},
         volume = {128},
         issn = {2169-9356},
         url = {https://onlinelibrary.wiley.com/doi/abs/10.1029/2022JB025610},
@@ -158,7 +158,7 @@
 }
 
 @misc{netcdf,
-  title={Network Common Data Form (NetCDF), version 4},
+  title={{N}etwork {C}ommon {D}ata {F}orm ({NetCDF}), version 4},
   publisher={UCAR/Unidata, Colder, CO, USA},
   author={Unidata},
   year=2008,
