### 2025-12-15 ee0a863
JOSS: Final fix for references
Correct error in Bunge et al. (1997).
