+++ paper/paper.bib
@@ -7,7 +7,7 @@
                   Timo Heister and
 		  Robert Myhill and
                   John Naliboff},
-  title        = {ASPECT v2.4.0},
+  title        = {{ASPECT} v2.4.0},
   month        = jul,
   year         = 2022,
   publisher    = {Zenodo},
@@ -30,7 +30,7 @@
 }
 
 @article{bunge1997,
-  title={A sensitivity study of three-dimensional spherical mantle convection at 108 Rayleigh number: Effects of depth-dependent viscosity, heating mode, and an endothermic phase change},
+  title={A sensitivity study of three-dimensional spherical mantle convection at 10$^8$ {R}ayleigh number: {E}ffects of depth-dependent viscosity, heating mode, and an endothermic phase change},
   author={Bunge, Hans-Peter and Richards, Mark A and Baumgardner, John R},
   journal={Journal of Geophysical Research: Solid Earth},
   volume={102},
