### 2025-12-15 #7539 @anowacki
@kthyng Sorry about that—one error slipped through after the [previous change](https://github.com/mantle-convection-constrained/terratools/commit/e1682ee0c077c6c8397ec485a070beee45d63dcf).  That was the final error that I missed, apologies.
