+++ .github/workflows/main.yml
@@ -24,7 +24,7 @@ jobs:
         ./contrib/utilities/indent
         git diff > changes-astyle.diff
     - name: archive indent results
-      uses: actions/upload-artifact@v5
+      uses: actions/upload-artifact@v6
       with:
         name: changes-astyle.diff
         path: changes-astyle.diff
